Fix unsigned int / NSUInteger mismatch.
git-svn-id: http://svn.gna.org/svn/etoile/trunk/Etoile/Frameworks/EtoileUI@7981 7b3f36aa-e6db-49c1-8b4c-66eac1997e64

+++ Source/NSWindow+Etoile.m
@@ -185,7 +185,7 @@ top left coordinates when the content view returns YES to -isFlipped.  */
 }
 
 - (id) initWithContentRect: (NSRect)contentRect 
-                 styleMask: (unsigned int)windowStyle
+                 styleMask: (NSUInteger)windowStyle
                    backing: (NSBackingStoreType)bufferingType 
                      defer: (BOOL)deferCreation
 {
